Splitting the camera into its own file
