+++ game/engine/headers/camera.h
@@ -1,0 +1,4 @@
+#include <cglm/cglm.h>
+
+void setCameraPos(vec3 cameraPos, vec3 cameraDir, int cameraUniformPos);
+void initCamera(int cameraUniformPos);
