+++ game/glFunctions/headers/keyboard.h
@@ -1,1 +1,2 @@
-void keyboard(unsigned char key, int x, int y);
+#include <GLFW/glfw3.h>
+void keyboard(GLFWwindow* window, int key, int scancode, int action, int mods);
